fix: settings page #11
